Allow x_band, x_lines, x_minrange functionality for any plot type (specifically Scatter)
