+++ multiqc/plots/plotly/line.py
@@ -10,7 +10,7 @@ from plotly.graph_objs.layout.shape import Label  # type: ignore
 from pydantic import BaseModel, Field
 
 from multiqc import config, report
-from multiqc.plots.plotly.plot import BaseDataset, PConfig, Plot, PlotType
+from multiqc.plots.plotly.plot import BaseDataset, PConfig, Plot, PlotType, convert_dash_style
 from multiqc.types import SampleName
 from multiqc.utils.util_functions import update_dict
 from multiqc.validation import ValidatedConfig, add_validation_warning
@@ -62,50 +62,17 @@ class Series(ValidatedConfig, Generic[KeyTV, ValueTV]):
 
         super().__init__(**data, _parent_class=LinePlotConfig)
 
+    def get_x_range(self) -> Tuple[Optional[KeyTV], Optional[KeyTV]]:
+        xs = [x[0] for x in self.pairs]
+        if len(xs) > 0:
+            return min(xs), max(xs)
+        return None, None
 
-class FlatLine(ValidatedConfig):
-    """
-    Extra X=const or Y=const line added to the plot
-    """
-
-    value: Union[float, int]
-    colour: Optional[str] = Field(None, deprecated="color")
-    color: Optional[str] = None
-    width: int = 2
-    dashStyle: Optional[str] = Field(None, deprecated="dash")
-    dash: Optional[str] = None
-    label: Optional[Union[str, Dict]] = None
-
-    @classmethod
-    def parse_label(cls, value):
-        if isinstance(value, dict):
-            add_validation_warning(
-                [LinePlotConfig, FlatLine],
-                "Line plot's x_lines or y_lines 'label' field is expected to be a string. "
-                "Other fields other than 'text' are deprecated and will be ignored",
-            )
-            return value["text"]
-        return value
-
-    def __init__(self, **data):
-        if "dashStyle" in data:
-            data["dash"] = convert_dash_style(FlatLine, data.pop("dashStyle"))
-        if "dash" in data:
-            data["dash"] = convert_dash_style(FlatLine, data["dash"])
-        super().__init__(**data, _parent_class=LinePlotConfig)
-
-
-class LineBand(ValidatedConfig):
-    """
-    Extra X1-X2 or Y1-Y2 band added to the plot
-    """
-
-    from_: Union[float, int]
-    to: Union[float, int]
-    color: Optional[str] = None
-
-    def __init__(self, **data):
-        super().__init__(**data, _parent_class=LinePlotConfig)
+    def get_y_range(self) -> Tuple[Optional[ValueTV], Optional[ValueTV]]:
+        ys = [x[1] for x in self.pairs if x[1] is not None]
+        if len(ys) > 0:
+            return min(ys), max(ys)
+        return None, None
 
 
 SeriesT = Union[Series, Dict]
@@ -118,18 +85,6 @@ class LinePlotConfig(PConfig):
     smooth_points: Optional[int] = 500
     smooth_points_sumcounts: Union[bool, List[bool], None] = None
     extra_series: Optional[Union[Series, List[Series], List[List[Series]]]] = None
-    xMinRange: Optional[Union[float, int]] = Field(None, deprecated="x_minrange")
-    yMinRange: Optional[Union[float, int]] = Field(None, deprecated="y_minrange")
-    x_minrange: Optional[Union[float, int]] = None
-    y_minrange: Optional[Union[float, int]] = None
-    xPlotBands: Optional[List[LineBand]] = Field(None, deprecated="x_bands")
-    yPlotBands: Optional[List[LineBand]] = Field(None, deprecated="y_bands")
-    xPlotLines: Optional[List[FlatLine]] = Field(None, deprecated="x_lines")
-    yPlotLines: Optional[List[FlatLine]] = Field(None, deprecated="y_lines")
-    x_bands: Optional[List[LineBand]] = None
-    y_bands: Optional[List[LineBand]] = None
-    x_lines: Optional[List[FlatLine]] = None
-    y_lines: Optional[List[FlatLine]] = None
     style: Optional[Literal["lines", "lines+markers"]] = None
     hide_zero_cats: Optional[bool] = Field(False, deprecated="hide_empty")
     hide_empty: bool = False
@@ -142,22 +97,6 @@ class LinePlotConfig(PConfig):
                 return [[Series(**d) if isinstance(d, dict) else d for d in ds] for ds in data]
             return [Series(**d) if isinstance(d, dict) else d for d in data]
         return Series(**data) if isinstance(data, dict) else data
-
-    @classmethod
-    def parse_x_bands(cls, data):
-        return [LineBand(**d) for d in data]
-
-    @classmethod
-    def parse_y_bands(cls, data):
-        return [LineBand(**d) for d in data]
-
-    @classmethod
-    def parse_x_lines(cls, data):
-        return [FlatLine(**d) for d in data]
-
-    @classmethod
-    def parse_y_lines(cls, data):
-        return [FlatLine(**d) for d in data]
 
 
 def plot(lists_of_lines: List[List[Series]], pconfig: LinePlotConfig) -> "LinePlot":
@@ -179,6 +118,30 @@ def plot(lists_of_lines: List[List[Series]], pconfig: LinePlotConfig) -> "LinePl
 
 class Dataset(BaseDataset):
     lines: List[Series]
+
+    def get_x_range(self) -> Tuple[Optional[KeyTV], Optional[KeyTV]]:
+        if not self.lines:
+            return None, None
+        xmax, xmin = None, None
+        for line in self.lines:
+            _xmin, _xmax = line.get_x_range()
+            if _xmin is not None:
+                xmin = _xmin if xmin is None else min(xmin, _xmin)
+            if _xmax is not None:
+                xmax = _xmax if xmax is None else max(xmax, _xmax)
+        return xmin, xmax
+
+    def get_y_range(self) -> Tuple[Optional[ValueTV], Optional[ValueTV]]:
+        if not self.lines:
+            return None, None
+        ymax, ymin = None, None
+        for line in self.lines:
+            _ymin, _ymax = line.get_y_range()
+            if _ymin is not None:
+                ymin = _ymin if ymin is None else min(ymin, _ymin)
+            if _ymax is not None:
+                ymax = _ymax if ymax is None else max(ymax, _ymax)
+        return ymin, ymax
 
     @staticmethod
     def create(
@@ -332,160 +295,7 @@ def create(
     # Make a tooltip always show on hover over any point on plot
     model.layout.hoverdistance = -1
 
-    y_minrange = pconfig.y_minrange
-    x_minrange = pconfig.x_minrange
-    y_bands = pconfig.y_bands
-    x_bands = pconfig.x_bands
-    x_lines = pconfig.x_lines
-    y_lines = pconfig.y_lines
-    if y_bands or y_lines or y_minrange:
-        # We don't want the bands to affect the calculated axis range, so we
-        # find the min and the max from data points, and manually set the range.
-        for dataset in model.datasets:
-            minval = dataset.layout["yaxis"]["autorangeoptions"]["minallowed"]
-            maxval = dataset.layout["yaxis"]["autorangeoptions"]["maxallowed"]
-            for line in dataset.lines:
-                ys = [x[1] for x in line.pairs]
-                if len(ys) > 0:
-                    minval = min(ys) if minval is None else min(minval, min(ys))
-                    maxval = max(ys) if maxval is None else max(maxval, max(ys))
-            if maxval is not None and minval is not None:
-                maxval += (maxval - minval) * 0.05
-            clipmin = dataset.layout["yaxis"]["autorangeoptions"]["clipmin"]
-            clipmax = dataset.layout["yaxis"]["autorangeoptions"]["clipmax"]
-            if clipmin is not None and minval is not None and clipmin > minval:
-                minval = clipmin
-            if clipmax is not None and maxval is not None and clipmax < maxval:
-                maxval = clipmax
-            if y_minrange is not None and maxval is not None and minval is not None:
-                maxval = max(maxval, minval + y_minrange)
-            if model.layout.yaxis.type == "log":
-                minval = math.log10(minval) if minval is not None and minval > 0 else None
-                maxval = math.log10(maxval) if maxval is not None and maxval > 0 else None
-            dataset.layout["yaxis"]["range"] = [minval, maxval]
-
-    if x_bands or x_lines or (x_minrange and not pconfig.categories):
-        # same as above but for x-axis
-        for dataset in model.datasets:
-            minval = dataset.layout["xaxis"]["autorangeoptions"]["minallowed"]
-            maxval = dataset.layout["xaxis"]["autorangeoptions"]["maxallowed"]
-            for series in dataset.lines:
-                xs = [x[0] for x in series.pairs]
-                if len(xs) > 0:
-                    minval = min(xs) if minval is None else min(minval, min(xs))
-                    maxval = max(xs) if maxval is None else max(maxval, max(xs))
-            clipmin = dataset.layout["xaxis"]["autorangeoptions"]["clipmin"]
-            clipmax = dataset.layout["xaxis"]["autorangeoptions"]["clipmax"]
-            if clipmin is not None and minval is not None and clipmin > minval:
-                minval = clipmin
-            if clipmax is not None and maxval is not None and clipmax < maxval:
-                maxval = clipmax
-            if x_minrange is not None and maxval is not None and minval is not None:
-                maxval = max(maxval, minval + x_minrange)
-            if model.layout.xaxis.type == "log":
-                minval = math.log10(minval) if minval is not None and minval > 0 else None
-                maxval = math.log10(maxval) if maxval is not None and maxval > 0 else None
-            dataset.layout["xaxis"]["range"] = [minval, maxval]
-
-    model.layout.shapes = (
-        [
-            dict(
-                type="rect",
-                y0=band.from_,
-                y1=band.to,
-                x0=0,
-                x1=1,
-                xref="paper",  # make x coords are relative to the plot paper [0,1]
-                fillcolor=band.color,
-                line={
-                    "width": 0,
-                },
-                layer="below",
-            )
-            for band in (y_bands or [])
-        ]
-        + [
-            dict(
-                type="rect",
-                x0=band.from_,
-                x1=band.to,
-                y0=0,
-                y1=1,
-                yref="paper",  # make y coords are relative to the plot paper [0,1]
-                fillcolor=band.color,
-                line={
-                    "width": 0,
-                },
-                layer="below",
-            )
-            for band in (x_bands or [])
-        ]
-        + [
-            dict(
-                type="line",
-                xref="paper",
-                yref="y",
-                x0=0,
-                y0=line.value,
-                x1=1,
-                y1=line.value,
-                line={
-                    "width": line.width,
-                    "dash": line.dash,
-                    "color": line.color,
-                },
-                label=dict(text=line.label, font=dict(color=line.color)),
-            )
-            for line in (y_lines or [])
-        ]
-        + [
-            dict(
-                type="line",
-                yref="paper",
-                xref="x",
-                x0=line.value,
-                y0=0,
-                x1=line.value,
-                y1=1,
-                line={
-                    "width": line.width,
-                    "dash": line.dash,
-                    "color": line.color,
-                },
-                label=dict(text=line.label, font=dict(color=line.color)),
-            )
-            for line in (x_lines or [])
-        ]
-    )
-
     return LinePlot(**model.__dict__)
-
-
-def convert_dash_style(cls: type, dash_style: Optional[str]) -> Optional[str]:
-    """Convert dash style from Highcharts to Plotly"""
-    if dash_style is None:
-        return None
-    mapping = {
-        "Solid": "solid",
-        "ShortDash": "dash",
-        "ShortDot": "dot",
-        "ShortDashDot": "dashdot",
-        "ShortDashDotDot": "dashdot",
-        "Dot": "dot",
-        "Dash": "dash",
-        "DashDot": "dashdot",
-        "LongDash": "longdash",
-        "LongDashDot": "longdashdot",
-        "LongDashDotDot": "longdashdot",
-    }
-    if dash_style in mapping.values():  # Plotly style?
-        return dash_style
-    elif dash_style in mapping.keys():  # Highcharts style?
-        add_validation_warning(
-            [LinePlotConfig, cls], f"'{dash_style}' is a deprecated dash style, use '{mapping[dash_style]}'"
-        )
-        return mapping[dash_style]
-    return "solid"
 
 
 def remove_nones_and_empty_dicts(d: Mapping) -> Dict:
